2014-08-09 19:14 UTC+0200 Viktor Szakats (vszakats users.noreply.github.com)   * extras/gtwvw/*     ! fixed two warnings after latest changes

+++ extras/gtwvw/gtwvwd.c
@@ -5743,7 +5743,7 @@ static void  hb_gt_wvw_vxPutch( WIN_DATA * pWindowData, USHORT iRow, USHORT iCol
    if( index < pWindowData->BUFFERSIZE )
    {
       pWindowData->pBuffer[ index ] = bChar;
-      pWindowData->pColors[ index ] = bColor;
+      pWindowData->pColors[ index ] = ( BYTE ) bColor;
 
       /* determine bounds of rect around character to refresh */
       hb_gt_wvw_SetInvalidRect( pWindowData, iCol, iRow, iCol, iRow );
